refactoring to 3.0 book

+++ src/camera.h
@@ -10,7 +10,7 @@ class camera {
     vertical = vec3(0.0, 2.0, 0.0);
     origin = vec3(0.0, 0.0, 0.0);
   }
-  ray get_ray(float u, float v) {
+  ray get_ray(double u, double v) {
     return ray(origin,
                lower_left_corner + u * horizontal + v * vertical - origin);
   }
